# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'lciii-scaffolding'","description":"'LCII Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'LCII Initial Scaffolding'"}],"paths":{"/api/countries":{"get":{"tags":["country-controller"],"operationId":"getAllCountries","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}},"post":{"tags":["country-controller"],"operationId":"saveRandomCountries","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveCountriesRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ResponseDTO"}}}}}}}},"/api/countries/{language}/language":{"get":{"tags":["country-controller"],"operationId":"getCountriesByLanguage","parameters":[{"name":"language","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"/api/countries/{continent}/continent":{"get":{"tags":["country-controller"],"operationId":"getCountriesByContinent","parameters":[{"name":"continent","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"/api/countries/{code}":{"get":{"tags":["country-controller"],"operationId":"getFilteredCountriesByCode","parameters":[{"name":"code","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"/api/countries/most-borders":{"get":{"tags":["country-controller"],"operationId":"getCountryWithMostBorders","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CountryDTO"}}}}}}},"/api/countries/":{"get":{"tags":["country-controller"],"operationId":"getFilteredCountriesByName","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}}},"components":{"schemas":{"SaveCountriesRequest":{"type":"object","properties":{"amountOfCountryToSave":{"type":"integer","format":"int32"}}},"ResponseDTO":{"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"}}},"CountryDTO":{"type":"object","properties":{"common":{"type":"string"},"official":{"type":"string"},"tld":{"type":"array","items":{"type":"string"}},"cca2":{"type":"string"},"ccn3":{"type":"string"},"cca3":{"type":"string"},"cioc":{"type":"string"},"independent":{"type":"boolean"},"unMember":{"type":"boolean"},"currencies":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/CurrencyDTO"}},"capital":{"type":"array","items":{"type":"string"}},"region":{"type":"string"},"subregion":{"type":"string"},"languages":{"type":"object","additionalProperties":{"type":"string"}},"borders":{"type":"object","additionalProperties":{"type":"string"}},"area":{"type":"number","format":"double"},"population":{"type":"integer","format":"int64"},"flag":{"type":"string"},"latlng":{"type":"array","items":{"type":"number","format":"double"}},"timezones":{"type":"array","items":{"type":"string"}}}},"CurrencyDTO":{"type":"object","properties":{"name":{"type":"string"},"symbol":{"type":"string"}}}}}}