# src/main/java/ar/edu/utn/frc/tup/lciii/DTO/ResponseDTO.java
package ar.edu.utn.frc.tup.lciii.DTO;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class ResponseDTO {
    private String code;
    private String name;
}